api/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
azure-storage-blob>=12.23.0,<13.0.0
azure-identity>=1.19.0,<2.0.0
pyodbc>=5.2.0,<6.0.0
azure-search-documents>=11.6.0,<12.0.0
workos>=7.0.0,<8.0.0
pydantic-settings>=2.6.0,<3.0.0
python-multipart>=0.0.18,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
httpx>=0.28.0,<1.0.0
